Add classes into files
